refactor: split pipes.rs
