+++ src/structure.rs
@@ -1,3 +1,5 @@
+mod pipes;
+
 use std::collections::{HashMap, VecDeque};
 
 use serde::{Deserialize, Serialize};
@@ -7,6 +9,10 @@ use crate::{
     train_tracks::{SegmentDirection, StationId, TrainTask},
     vec2::Vec2,
 };
+
+use self::pipes::{PIPE_FLOW_RATE, WATER_PUMP_RATE};
+
+pub(crate) use self::pipes::{FluidBox, FluidType, MAX_FLUID_AMOUNT, Pipe, PipeId};
 
 use std::f64::consts::PI;
 
@@ -28,9 +34,6 @@ pub(crate) const MAX_BELT_LENGTH: f64 = 20.;
 pub(crate) const ITEM_INTERVAL: f64 = 1.0;
 pub(crate) const BELT_SPEED: f64 = 0.05; // length per tick
 pub(crate) const BELT_MAX_SLOPE: f64 = 0.1;
-pub(crate) const MAX_FLUID_AMOUNT: f64 = 100.;
-const WATER_PUMP_RATE: f64 = 0.1;
-const PIPE_FLOW_RATE: f64 = 0.2;
 
 pub(crate) type StructureId = usize;
 
@@ -912,102 +915,6 @@ fn normalize_car_idx(train: &Train, car_idx: i32) -> i32 {
     }
 }
 
-pub(crate) type PipeId = usize;
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub(crate) struct Pipe {
-    pub(crate) start: Vec2<f64>,
-    pub(crate) start_con: BeltConnection,
-    pub(crate) end: Vec2<f64>,
-    pub(crate) end_con: BeltConnection,
-    pub(crate) fluid: Option<FluidBox>,
-}
-
-impl Pipe {
-    pub fn new(
-        start: Vec2<f64>,
-        start_con: BeltConnection,
-        end: Vec2<f64>,
-        end_con: BeltConnection,
-    ) -> Self {
-        Self {
-            start,
-            start_con,
-            end,
-            end_con,
-            fluid: None,
-        }
-    }
-
-    fn update(&mut self) -> PipeUpdateResult {
-        let mut ret = PipeUpdateResult::default();
-        if let Some(fluid) = &mut self.fluid {
-            let fluid_box = FluidBox {
-                amount: (fluid.amount * PIPE_FLOW_RATE)
-                    .max(PIPE_FLOW_RATE)
-                    .min(fluid.amount),
-                ty: fluid.ty,
-            };
-            let fullness = fluid.amount / MAX_FLUID_AMOUNT;
-            for con in [self.start_con, self.end_con] {
-                match con {
-                    BeltConnection::BeltStart(pipe_id) | BeltConnection::BeltEnd(pipe_id) => {
-                        ret.moved_fluids
-                            .push((EntityId::Pipe(pipe_id), fluid_box, fullness));
-                    }
-                    BeltConnection::Structure(st_id, _) => {
-                        ret.moved_fluids
-                            .push((EntityId::Structure(st_id), fluid_box, fullness));
-                    }
-                    _ => {}
-                }
-            }
-        }
-        ret
-    }
-
-    /// Try to insert some amount of fluid. The rate of flow depends on the pressure and amount already in the pipe.
-    /// If the pipe has another type of fluid, it will be rejected.
-    /// Returns the amount of fluid actually moved.
-    fn try_insert(&mut self, incoming_fluid: FluidBox, pressure: f64) -> Option<FluidBox> {
-        let Some(fluid_box) = &mut self.fluid else {
-            self.fluid = Some(incoming_fluid);
-            return Some(incoming_fluid);
-        };
-        if fluid_box.ty == incoming_fluid.ty && fluid_box.amount < MAX_FLUID_AMOUNT {
-            let my_pressure = fluid_box.amount / MAX_FLUID_AMOUNT;
-            let delta = pressure - my_pressure;
-            if delta < 0. {
-                return None;
-            }
-            let flow = delta * PIPE_FLOW_RATE;
-            let after = (fluid_box.amount + flow).min(MAX_FLUID_AMOUNT);
-            let ret = FluidBox {
-                amount: after - fluid_box.amount,
-                ty: fluid_box.ty,
-            };
-            fluid_box.amount = after;
-            return Some(ret);
-        }
-        None
-    }
-
-    /// Attempt to remove items that were successfully deleted.
-    pub fn post_update(&mut self, fluid: FluidBox) {
-        if let Some(fluid_box) = &mut self.fluid {
-            fluid_box.amount = (fluid_box.amount - fluid.amount).max(0.0);
-            if fluid_box.amount == 0. {
-                self.fluid = None;
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug, Default)]
-struct PipeUpdateResult {
-    moved_fluids: Vec<(EntityId, FluidBox, f64)>,
-}
-
 #[derive(Clone, Copy, Debug)]
 pub(crate) struct OreVein {
     pub pos: Vec2,
@@ -1057,16 +964,4 @@ impl Inventory {
         self.coal += count;
         self
     }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) enum FluidType {
-    Water,
-    Steam,
-}
-
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
-pub(crate) struct FluidBox {
-    pub amount: f64,
-    pub ty: FluidType,
 }
